Modularize the whole project (#8)
This commit tidies up the project and divides it into modular packages.
