+++ application/app/src/main/java/com/example/csi_dmce/MainActivity.kt
@@ -3,6 +3,7 @@ package com.example.csi_dmce
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.csi_dmce.ui.Dashboard
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
