+++ application/app/src/main/java/com/example/csi_dmce/auth/ForgotPassword.kt
@@ -1,10 +1,11 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.auth
 
 import android.os.Bundle
 import android.widget.Button
 import android.widget.EditText
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import com.example.csi_dmce.R
 
 class ForgotPassword: AppCompatActivity() {
     private lateinit var tv_user_phone_number: TextView
