+++ application/app/src/main/java/com/example/csi_dmce/calendar/CSICalendar.kt
@@ -1,7 +1,8 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.calendar
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.csi_dmce.R
 
 class CSICalendar: AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
