+++ application/app/src/main/java/com/example/csi_dmce/database/AppDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.database
 
 import android.content.Context
 import androidx.room.Database
