+++ application/app/src/main/java/com/example/csi_dmce/database/User.kt
@@ -1,4 +1,4 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
