+++ application/app/src/main/java/com/example/csi_dmce/events/Event_page.kt
@@ -1,11 +1,12 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.events
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.csi_dmce.R
 
-class event_items : AppCompatActivity() {
+class event_page : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_event_items)
+        setContentView(R.layout.activity_event_page)
     }
 }
