+++ application/app/src/main/java/com/example/csi_dmce/events/Events.kt
@@ -1,4 +1,4 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.events
 
 data class events(var titleimage : Int, var eventname : String)
