+++ application/app/src/main/java/com/example/csi_dmce/profile/Profile.kt
@@ -1,4 +1,4 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.profile
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -8,6 +8,8 @@ import android.widget.Button
 import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.csi_dmce.profile.ProfileEvents
+import com.example.csi_dmce.profile.ProfileEventsAdapter
 
 private  lateinit var newRecyclerView: RecyclerView
 private  lateinit var newArrayList: ArrayList<ProfileEvents>
