+++ application/app/src/main/java/com/example/csi_dmce/profile/ProfileEvents.kt
@@ -1,4 +1,4 @@
-package com.example.csi_dmce
+package com.example.csi_dmce.profile
 
 data class ProfileEvents(var title_heading : String, var Description : String)
