+++ application/app/src/main/res/layout/activity_event_items.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".event_items"
+    tools:context=".events.event_items"
     android:layout_margin="10dp">
 
     <com.google.android.material.imageview.ShapeableImageView
