+++ application/app/src/main/res/layout/activity_event_recycler.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".event_recycler">
+    tools:context=".events.event_recycler">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
